app/db.py
import os
from dotenv import load_dotenv
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker, declarative_base

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
# sync engine
engine = create_engine(DATABASE_URL, pool_pre_ping=True)

# класс фабрики сессий
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

# базовый класс для моделей
Base = declarative_base()
